refactor(networking): Isolate APIError enum into its own file
